backend/requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
Pillow==11.3.0
boto3==1.39.14
python-decouple==3.8